% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minifyCSS.R
\name{minifyCSS}
\alias{minifyCSS}
\title{CSS Minifier}
\usage{
minifyCSS(css)
}
\arguments{
\item{css}{Character string containing CSS information}
}
\description{
This function uses csso. It performs three sort of transformations: cleaning (removing redundant),
 compression (replacement for shorter form) and restructuring (merge of declarations, rulesets and so on).
 As a result your CSS becomes much smaller.
}
\examples{
input = "* { color: green; }
ul, ol, li { color: blue; }
UL.foo, span.bar { color: red; }
"
minifyCSS(input)
}
\seealso{
https://github.com/css/csso
}
